compat fixes
