+++ addons/compat_rf/CfgMainMenuSpotlight.hpp
@@ -3,4 +3,5 @@ class CfgMainMenuSpotlight
 	delete Scenario_AirControl_RF;
     delete Scenario_FireSeason_RF;
     delete Scenario_TrainingDay_RF;
+	
 };
