+++ addons/compat_rf/config.cpp
@@ -12,7 +12,7 @@ class CfgPatches {
         requiredVersion = REQUIRED_VERSION; 
         // Required addons, used for setting load order.
         // When any of the addons is missing, pop-up warning will appear when launching the game.
-        requiredAddons[] = { "ghg_main", "data_rf", "missions_rf"};
+        requiredAddons[] = { "ghg_main", "RF_Data", "RF_Missions"};
         // List of objects (CfgVehicles classes) contained in the addon. Important also for Zeus content (units and groups) unlocking.
         units[] = {};
         // List of weapons (CfgWeapons classes) contained in the addon.
