+++ addons/compat_rhs/XEH_PREP.hpp
@@ -1,0 +1,3 @@
+PREP(c130CanDrop);
+PREP(c130DropPlane);
+PREP(c130DropPayload);
