+++ addons/compat_rhs/functions/fn_cargoAttach.sqf
