+++ addons/compat_rhs/functions/fnc_c130DropPayload.sqf
