+++ addons/compat_sog/XEH_PREP.hpp
@@ -1,2 +1,7 @@
 PREP(fixHueyHud);
 PREP(registerJukeboxTracks);
+PREP(anestheticInit);
+PREP(anestheticHandler);
+PREP(anestheticWoundHandler);
+PREP(blu82Init);
+PREP(blu82Deploy);
