+++ addons/compat_sog/XEH_postInit.sqf
@@ -3,3 +3,5 @@
     Everything in this file is run once on mission start
 	Author: CBA_3
 ======================================*/
+[] call FUNC(anestheticInit);
+[] call FUNC(blu82Init);
