+++ addons/compat_sog/functions/fn_blu82Hit.sqf
