+++ addons/compat_sog/functions/fnc_anestheticHandler.sqf
