+++ addons/compat_sog/functions/fnc_anestheticInit.sqf
