+++ addons/compat_sog/functions/fnc_anestheticWoundHandler.sqf
