+++ addons/compat_sog/functions/fnc_blu82Deploy.sqf
