+++ addons/compat_sog/functions/fnc_blu82Init.sqf
