+++ addons/compat_sog/models/dart/dart.p3d
