+++ addons/specweap/XEH_PREP.hpp
@@ -1,8 +1,1 @@
-PREP(anestheticInit);
-PREP(anestheticHandler);
-PREP(anestheticWoundHandler);
-PREP(blu82Init);
-PREP(blu82Deploy);
-PREP(c130CanDrop);
-PREP(c130DropPlane);
-PREP(c130DropPayload);
+
