+++ addons/specweap/XEH_postInit.sqf
@@ -4,5 +4,4 @@ if ! ( GVARMAIN(is_mission_ghg) ) exitWith {};
     Everything in this file is run once on mission start
 	Author: CBA_3
 ======================================*/
-[] call FUNC(anestheticInit);
-[] call FUNC(blu82Init);
+
